Update Addgames.html
Updated header and footer. Previous commit seems to not have worked because... did I just updated it in the wrong branch? That must not be posible, because there is only one branch in the current repository. Anyway, I don't know how it fail to upload the files, and replace the previous.